PythonQAOtus_FinalProject/requirements.txt
pytest==5.4.3
selenium==3.141.0
pytest-xdist==1.32.0
allure-pytest==2.7.0
requests==2.24.0
